+++ packages/iov-bns/types/tags.d.ts
@@ -1,3 +1,3 @@
-import { Address, BcpQueryTag, BcpSwapQuery } from "@iov/bcp-types";
+import { Address, AtomicSwapQuery, BcpQueryTag } from "@iov/bcp-types";
 export declare function bnsNonceTag(addr: Address): BcpQueryTag;
-export declare function bnsSwapQueryTags(query: BcpSwapQuery, set?: boolean): BcpQueryTag;
+export declare function bnsSwapQueryTag(query: AtomicSwapQuery, set?: boolean): BcpQueryTag;